Move entry points to ./lib/exports
Inspired by SvelteKit (https://github.com/sveltejs/kit)

Motivation:
- the entry points are indeed part of the library source code
- simplifies lint:javascript script and files field in package.json
- elimintates the need for "rootDir": ".." in TSConfig for lib
  (there are no more typings emitted for untracked JS files in the root
  directory, for example config files and files with quick tests/demos)
