+++ lib/exports/index.js
@@ -1,8 +1,8 @@
-const { Argument } = require('./lib/argument.js');
-const { Command } = require('./lib/command.js');
-const { CommanderError, InvalidArgumentError } = require('./lib/error.js');
-const { Help } = require('./lib/help.js');
-const { Option } = require('./lib/option.js');
+const { Argument } = require('../argument.js');
+const { Command } = require('../command.js');
+const { CommanderError, InvalidArgumentError } = require('../error.js');
+const { Help } = require('../help.js');
+const { Option } = require('../option.js');
 
 /**
  * Expose the root command.
